Fixed the threshold functionalities in legend and gene plotter

+++ src/viewOpenGL/GeneData.h
@@ -29,14 +29,14 @@ public:
     void updateQuadColor(const int index, QColor4ub newcolor);
     void updateQuadVisible(const int index, bool visible);
     void updateQuadSelected(const int index, bool selected);
-    void updateQuadRefCount(const int index, float refcount);
-    void updateQuadValue(const int index, float value);
+    void updateQuadRefCount(const int index, qreal refcount);
+    void updateQuadValue(const int index, qreal value);
 
     QColor4ub quadColor(const int index) const;
     bool quadVisible(const int index) const;
     bool quadSelected(const int index) const;
-    float quadRefCount(const int index) const;
-    float quadValue(const int index) const;
+    qreal quadRefCount(const int index) const;
+    qreal quadValue(const int index) const;
 
     void resetRefCount();
     void resetValues();
